requirements.txt
python-dotenv
azure-identity
azure-ai-projects
agent-framework-azure-ai
fastapi
uvicorn